requirements.txt
tensorflow
pillow
numpy
pandas
flask
prometheus_client